# packages/backend/src/infrastructure/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Equipment Rental System API
  description: |
    A comprehensive API for managing equipment rentals, reservations, and members.
    Built using hexagonal architecture with TypeScript.

    ## Features
    - Equipment management and availability tracking
    - Member registration and tier-based benefits
    - Rental creation, extension, and return workflows
    - Reservation system with payment authorization
    - Damage assessment and late fee calculation
    - Payment processing integration
    - Notification system

    ## Architecture
    This API follows hexagonal (ports and adapters) architecture principles,
    ensuring clean separation of concerns and testability.
  version: 1.0.0
  contact:
    name: Equipment Rental System
    email: support@equipmentrental.com
servers:
  - url: http://localhost:3000
    description: Development server
  - url: http://localhost:3000/api
    description: API base path
...
paths:
...
  /api/equipment:
    get:
      summary: Get available equipment
      description: Retrieves a list of all equipment currently available for rent
      tags:
        - Equipment
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
            format: date
          description: Start date for availability check (ISO 8601)
        - in: query
          name: endDate
          schema:
            type: string
            format: date
          description: End date for availability check (ISO 8601)
      responses:
        '200':
          description: List of available equipment
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AvailableEquipment'
              example:
                - equipmentId: eq_123
                  name: Professional Camera
                  description: High-end DSLR camera
                  category: Photography
                  dailyRate: 50.00
                  condition: EXCELLENT
                - equipmentId: eq_456
                  name: Mountain Bike
                  description: 21-speed mountain bike
                  category: Sports
                  dailyRate: 25.00
                  condition: GOOD
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/equipment/{equipmentId}:
    get:
      summary: Get equipment by ID
      description: Retrieves detailed information about a specific equipment item
      tags:
        - Equipment
      parameters:
        - $ref: '#/components/parameters/EquipmentId'
      responses:
        '200':
          description: Equipment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipment'
              example:
                equipmentId: eq_123
                name: Professional Camera
                description: High-end DSLR camera
                category: Photography
                dailyRate: 50.00
                condition: EXCELLENT
                isAvailable: true
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/members/{memberId}:
    get:
      summary: Get member by ID
      description: Retrieves detailed information about a specific member
      tags:
        - Members
      parameters:
        - $ref: '#/components/parameters/MemberId'
      responses:
        '200':
          description: Member details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
              example:
                memberId: mem_789
                name: John Doe
                email: john.doe@example.com
                tier: GOLD
                joinDate: '2024-01-01T00:00:00.000Z'
                isActive: true
                activeRentalCount: 2
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/members/{memberId}/rentals:
    get:
      summary: Get member rentals
      description: Retrieves all rentals for a specific member
      tags:
        - Members
      parameters:
        - $ref: '#/components/parameters/MemberId'
      responses:
        '200':
          description: List of member rentals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemberRental'
              example:
                - rentalId: rent_001
                  equipmentId: eq_123
                  equipmentName: Professional Camera
                  startDate: '2024-01-10T00:00:00.000Z'
                  endDate: '2024-01-15T00:00:00.000Z'
                  status: ACTIVE
                  totalCost: 250.00
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/rentals/member/{memberId}:
    get:
      summary: Get rentals by member
      description: Retrieves all rentals for a specific member (alias endpoint)
      tags:
        - Rentals
      parameters:
        - $ref: '#/components/parameters/MemberId'
      responses:
        '200':
          description: List of member rentals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MemberRental'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    EquipmentId:
      name: equipmentId
      in: path
      required: true
      description: Unique identifier for the equipment
      schema:
        type: string
        example: eq_123
    MemberId:
      name: memberId
      in: path
      required: true
      description: Unique identifier for the member
      schema:
        type: string
        example: mem_789
...
  schemas:
...
    AvailableEquipment:
      type: object
      required:
        - equipmentId
        - name
        - description
        - category
        - dailyRate
        - condition
      properties:
        equipmentId:
          type: string
          description: Unique identifier for the equipment
        name:
          type: string
          description: Name of the equipment
        description:
          type: string
          description: Detailed description
        category:
          type: string
          description: Equipment category
        dailyRate:
          type: number
          format: float
          description: Daily rental rate in dollars
        condition:
          type: string
          enum: [EXCELLENT, GOOD, FAIR, POOR, NEEDS_MAINTENANCE, RETIRED]
          description: Current condition of the equipment
    Equipment:
      allOf:
        - $ref: '#/components/schemas/AvailableEquipment'
        - type: object
          required:
            - isAvailable
          properties:
            isAvailable:
              type: boolean
              description: Whether equipment is currently available
            currentRentalId:
              type: string
              description: ID of current rental if equipment is rented

    # Member Schemas
    Member:
      type: object
      required:
        - memberId
        - name
        - email
        - tier
        - joinDate
        - isActive
        - activeRentalCount
      properties:
        memberId:
          type: string
          description: Unique identifier for the member
        name:
          type: string
          description: Full name of the member
        email:
          type: string
          format: email
          description: Email address
        tier:
          type: string
          enum: [BASIC, SILVER, GOLD, PLATINUM]
          description: Membership tier level
        joinDate:
          type: string
          format: date-time
          description: Date when member joined
        isActive:
          type: boolean
          description: Whether member account is active
        activeRentalCount:
          type: number
          description: Number of active rentals
    MemberRental:
      type: object
      required:
        - rentalId
        - equipmentId
        - equipmentName
        - startDate
        - endDate
        - status
        - totalCost
      properties:
        rentalId:
          type: string
          description: Unique identifier for the rental
        equipmentId:
          type: string
          description: ID of rented equipment
        equipmentName:
          type: string
          description: Name of rented equipment
        startDate:
          type: string
          format: date-time
          description: Rental start date
        endDate:
          type: string
          format: date-time
          description: Rental end date
        status:
          type: string
          enum: [ACTIVE, RETURNED, OVERDUE, CANCELLED]
          description: Current rental status
        totalCost:
          type: number
          format: float
          description: Total cost of rental

    # Rental Schemas
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              description: Additional error details
              additionalProperties: true
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            equipmentNotFound:
              value:
                error:
                  code: EQUIPMENT_NOT_FOUND
                  message: Equipment not found
                  details:
                    equipmentId: eq_123
            memberNotFound:
              value:
                error:
                  code: MEMBER_NOT_FOUND
                  message: Member not found
                  details:
                    memberId: mem_789
            rentalNotFound:
              value:
                error:
                  code: RENTAL_NOT_FOUND
                  message: Rental not found
                  details:
                    rentalId: rent_001
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: INTERNAL_SERVER_ERROR
              message: An unexpected error occurred